agents/livekit-hedra-avatar/requirements.txt
livekit-agents[hedra,openai,silero]~=1.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0

